+++ schema.sql
@@ -584,6 +584,7 @@ CREATE TABLE google_configs (
     "preDeliveryChecklistTemplateUrl" TEXT,
     "bizRegCertUrl" TEXT,
     "bankbookCopyUrl" TEXT,
+    "defaultRootFolderId" TEXT,
     "createdAt" TEXT NOT NULL,
     "updatedAt" TEXT NOT NULL
 );